✏️ [core,infra] Implement streamlabs websocket API and tests

+++ Core/Streamlabs/Events/IStreamlabsDonation.cs
@@ -7,7 +7,7 @@ public interface IStreamlabsDonation : IDonation {
   public string Message { get; protected set; }
   public string UserId { get; protected set; }
   public string Currency { get; protected set; }
-  public double Amount { get; protected set; }
+  public decimal Amount { get; protected set; }
   public string FormattedAmount { get; protected set; }
   public string FromName { get; protected set; }
 }
